of: get of_root directly in place of of_find_node_by_path("/")
There are a couple places in drivers/of where the root node is found
with a of_find_node_by_path("/") call. This call just returns
of_node_get(of_root) when the the path "/" is passed as a parameter.
So, lets fixup these instances to just do that instead.

Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>