ibmvfc: fix potential race with request purging and link down
During a hard reset there is a small window where we can leak an IO
between moving everything off the sent list to the purge list and the
link down where new scsi requests actually get blocked. This is a by
product of relaxing the locking around queuecommand.

Reorder these two operations to ensure new requests are blocked prior to
purging outstanding requests. This is only an issue in the hard reset
path since the CRQ is possibly still active. However, for conistency
also update this behavior in the transport event patch where the client
gets notified that the CRQ is disconnected.

Fixes: 654080d02edb ("scsi: ibmvfc: Relax locking around ibmvfc_queuecommand()")
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>