ibmvfc: hold list lock during npiv_logout_done sanity check
When completing the npiv logout response the sent list is tested to
ensure it is empty. This is a sanity check to ensure the VIOS completed
or cancelled all outstanding requests in response to the npiv logout
management request. The queue request lists are meant to be protected by
their individual list locks which is not done here.

Ensure that the list_empty check is protected by the is lock. It
suffices to do this protection with simple spin_(un)lock calls as
management completions are guaranteed to be called with irqs disabled
and host_lock held.

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>