ibmvfc: sanity check sub-queues after NPIV logout when channelized
Currently, the NPIV logout response handler only sanity checks the CRQ
to ensure all outstanding commands were completed. When MQ is enabled
and the host is channelized it is the sub-queues that need to be sanity
checked, and not the CRQ.

When mq_enabled and using_channels flags are set loop through the
sub-queues instead of CRQ to ensure request sent lists are empty.

Fixes: 9000cb998bcf ("scsi: ibmvfc: Enable MQ and set reasonable defaults")
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>