ibmvfc: Add logic for protocol specific fabric login
Prior to the addition of NVMeoF support NPIV login was all that was
required of the client for the VIOS to setup the connection to the
Fabric. The new interface introduces protocol specific Fabric login
MAD's that must be used when SUPPORT_SCSI or SUPPORT_NVMEOF capabilities
are set by the VIOS after NPIV login. These capabilities will only be
advertised by a VIOS that supports them, and only when a client has
requested the new interface features by setting either YES_SCSI or
YES_NVMEOF in its NPIV login request capabilities field.

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>